Optimize arp and genny note sorting
arp/genny do a full sort of all notes whenever a note is played or ends, so when more than one note event is emitted by another module in a single DSP block you can get multiple full sorts, which can cause the DSP thread to miss a block.
Instead of sorting the array each time the currently current notes changes, a flag is set to sort them later.  When arp/genny are about to compute a new note, the flag is checked, and the sorting is done only if needed.